# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Event API",
    "version": "1.0.0",
    "description": "This API allows users to manage events and address events.",
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
    "/events": {
      "get": {
        "summary": "Retrieve all events",
        "description": "This endpoint retrieves a list of all available events. If no events are found, a 404 response is returned. This endpoint does not require authentication.",
        "operationId": "getAllEvents",
        "tags": [
          "Events"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved a list of events.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No events found.",
            "content": {
              "application/json": {
                "examples": {
                  "noEvents": {
                    "summary": "No events available",
                    "value": {
                      "message": "No events found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred on the server while fetching the events.",
            "content": {
              "application/json": {
                "examples": {
                  "serverError": {
                    "summary": "Unexpected error",
                    "value": {
                      "message": "Error fetching events"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new event",
        "description": "This endpoint allows users to create a new event. The event requires basic information such as name, description, start and end times, price, and an optional image file. The user must be authenticated, and the request should include a valid bearer token in the Authorization header. During the process, the user's type will be automatically updated to 'Anunciante' if they are not already one.",
        "operationId": "createEvent",
        "consumes": [
          "multipart/form-data"
        ],
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "The data required to create a new event. All fields except the event image are mandatory.",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The name of the event. This should be a concise and descriptive title.",
                    "example": "Tech Conference 2024"
                  },
                  "description": {
                    "type": "string",
                    "description": "A detailed description of the event, outlining its purpose, agenda, or any other relevant information.",
                    "example": "A cutting-edge conference on emerging tech trends."
                  },
                  "startAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "The start date and time of the event in ISO 8601 format (e.g., YYYY-MM-DDTHH:mm:ssZ).",
                    "example": "2024-05-15T09:00:00Z"
                  },
                  "endAt": {
                    "type": "string",
                    "format": "date-time",
                    "description": "The end date and time of the event in ISO 8601 format (e.g., YYYY-MM-DDTHH:mm:ssZ).",
                    "example": "2024-05-16T17:00:00Z"
                  },
                  "price": {
                    "type": "number",
                    "format": "float",
                    "description": "The price of the event. This value should be a positive number, representing the cost in the applicable currency.",
                    "example": 199.99
                  },
                  "eventPicture": {
                    "type": "string",
                    "format": "binary",
                    "description": "An optional image file representing the event. Accepted file formats include PNG, JPEG, and JPG."
                  }
                },
                "required": [
                  "name",
                  "description",
                  "startAt",
                  "endAt",
                  "price"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The event was successfully created. The response includes the details of the newly created event.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                },
                "examples": {
                  "success": {
                    "summary": "Successful creation",
                    "value": {
                      "id": "123e4567-e89b-12d3-a456-426614174000",
                      "name": "Tech Conference 2024",
                      "description": "A cutting-edge conference on emerging tech trends.",
                      "startAt": "2024-05-15T09:00:00Z",
                      "endAt": "2024-05-16T17:00:00Z",
                      "price": 199.99,
                      "eventPicture": "/uploads/event_pictures/tech_conf_2024.png",
                      "userId": "123e4567-e89b-12d3-a456-426614174001",
                      "eventstatus_id": "456e4567-e89b-12d3-a456-426614174002"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - The request is invalid due to missing required fields or an invalid file format.",
            "content": {
              "application/json": {
                "examples": {
                  "missingFields": {
                    "summary": "Missing required fields",
                    "value": {
                      "message": "Missing required fields"
                    }
                  },
                  "invalidFileFormat": {
                    "summary": "Invalid file format",
                    "value": {
                      "message": "Invalid file type. Only PNG, JPEG, and JPG are allowed."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred on the server while processing the request.",
            "content": {
              "application/json": {
                "examples": {
                  "serverError": {
                    "summary": "Unexpected error",
                    "value": {
                      "message": "Error creating event"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/events/my": {
      "get": {
        "summary": "Get the events of the current user",
        "description": "This endpoint retrieves all events created by the authenticated user. The request requires a valid bearer token for authentication.",
        "operationId": "getUserEvents",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "Events"
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the user's events.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Event"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing token.",
            "content": {
              "application/json": {
                "examples": {
                  "unauthorized": {
                    "summary": "Invalid token",
                    "value": {
                      "message": "Unauthorized - Missing or invalid token"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No events found for the user.",
            "content": {
              "application/json": {
                "examples": {
                  "noEvents": {
                    "summary": "User has no events",
                    "value": {
                      "message": "No events found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred on the server while fetching the user's events.",
            "content": {
              "application/json": {
                "examples": {
                  "serverError": {
                    "summary": "Unexpected error",
                    "value": {
                      "message": "Error fetching events"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/events/{id}/status": {
      "put": {
        "summary": "Approve a specific event by ID",
        "description": "Allows an admin to approve an event by updating its status to 'Aprovado'. The event must have a current status of 'Pendente'.",
        "operationId": "updateEventStatus",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the event to approve.",
            "schema": {
              "type": "string",
              "example": "123e4567-e89b-12d3-a456-426614174000"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event status updated successfully to 'Approved'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Event status is not 'Pending'.",
            "content": {
              "application/json": {
                "examples": {
                  "notPending": {
                    "summary": "Invalid status",
                    "value": {
                      "message": "Event status is not pending"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Event or required status not found.",
            "content": {
              "application/json": {
                "examples": {
                  "eventNotFound": {
                    "summary": "Event not found",
                    "value": {
                      "message": "Event not found"
                    }
                  },
                  "statusNotFound": {
                    "summary": "Status not found",
                    "value": {
                      "message": "Event status not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred while updating the event status.",
            "content": {
              "application/json": {
                "examples": {
                  "serverError": {
                    "summary": "Unexpected error",
                    "value": {
                      "message": "Error updating event status"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/events/{id}/cancel": {
      "put": {
        "summary": "Cancel a specific event by ID",
        "description": "Allows the cancellation of a specific event by updating its status to 'Cancelado' and processing associated refunds for ticket purchases.",
        "operationId": "cancelEvent",
        "tags": [
          "Events"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the event to cancel.",
            "schema": {
              "type": "string",
              "example": "123e4567-e89b-12d3-a456-426614174000"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event cancelled successfully. Refunds for associated tickets have been processed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                },
                "examples": {
                  "success": {
                    "summary": "Event cancelled successfully",
                    "value": {
                      "eventID": "123e4567-e89b-12d3-a456-426614174000",
                      "status": "Cancelled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Event not found or required status not found.",
            "content": {
              "application/json": {
                "examples": {
                  "eventNotFound": {
                    "summary": "Event not found",
                    "value": {
                      "message": "Event not found"
                    }
                  },
                  "statusNotFound": {
                    "summary": "Status not found",
                    "value": {
                      "message": "Event status not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected error occurred while processing the cancellation.",
            "content": {
              "application/json": {
                "examples": {
                  "serverError": {
                    "summary": "Unexpected error",
                    "value": {
                      "message": "Error updating event status"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Bearer token authentication"
      }
    },
    "schemas": {
      "Event": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID of the event"
          },
          "name": {
            "type": "string",
            "description": "Name of the event"
          },
          "description": {
            "type": "string",
            "description": "Description of the event"
          },
          "startAt": {
            "type": "string",
            "format": "date-time",
            "description": "Start date and time of the event"
          },
          "endAt": {
            "type": "string",
            "format": "date-time",
            "description": "End date and time of the event"
          },
          "price": {
            "type": "number",
            "description": "Price of the event"
          },
          "eventPicture": {
            "type": "string",
            "description": "Event status ID"
          }
        }
      },
...
    }
  }
}